requirements.txt
feedparser==6.0.11
psycopg2-binary==2.9.9
python-dotenv==1.0.1
sgmllib3k==1.0.0
